requirements.txt
llama-index
google-generativeai
llama-index-llms-gemini
pdfplumber
python-dotenv
IPython
llama-index-embeddings-gemini
streamlit


-e .